Compatibility with the QT example app and tests
All tests have been passed successfully.
