+++ examples/qt/chatWindow.cpp
@@ -45,7 +45,7 @@ ChatWindow::~ChatWindow()
     GUI_LOG_DEBUG("Destroying chat window for chat %s", Id(mRoom.chatid()).toString().c_str());
     if (mUpdateSeenTimer)
     {
-        cancelTimeout(mUpdateSeenTimer);
+        cancelTimeout(mUpdateSeenTimer, NULL);
         mUpdateSeenTimer = 0;
     }
 }
@@ -204,7 +204,7 @@ void ChatWindow::onMembersBtn(bool)
         menu.adjustSize();
         menu.exec(ui.mMembersBtn->mapToGlobal(
             QPoint(-menu.width()+ui.mMembersBtn->width(), ui.mMembersBtn->height())));
-    });
+    }, NULL);
 }
 void ChatWindow::dropEvent(QDropEvent* event)
 {
@@ -324,7 +324,7 @@ void ChatWindow::showCantEditNotice(const QString& action)
     WaitMsg tooltip(*this);
     tooltip.addMsg(tr("Can't %1 - message is too old").arg(action));
     setTimeout([tooltip]()
-    {}, 2000);
+    {}, 2000, NULL);
 }
 
 void ChatWindow::onUnsentEditLoaded(chatd::Message& editmsg, bool oriMsgIsSending)
@@ -509,7 +509,7 @@ void MessageWidget::msgDeleted()
     a->setDuration(300);
     a->setEasingCurve(QEasingCurve::Linear);
     a->start(QAbstractAnimation::DeleteWhenStopped);
-    setTimeout([this]() { removeFromList(); }, 300);
+    setTimeout([this]() { removeFromList(); }, 300, NULL);
 }
 
 void MessageWidget::removeFromList()
