+++ src/karereCommon.cpp
@@ -2,11 +2,36 @@
 #include "stringUtils.h"
 #include "rtcModule/IRtcModule.h"
 #include "sdkApi.h"
+#include "base/timers.hpp"
+#include "megachatapi_impl.h"
+
+#ifdef USE_LIBWEBSOCKETS
+#include "waiter/libuvWaiter.h"
+#else
+#include "waiter/libeventWaiter.h"
+#endif
 
 namespace karere
 {
 const char* gDbSchemaVersionSuffix = "2";
 bool gCatchException = true;
+
+void globalInit(void(*postFunc)(void*, void*), uint32_t options, const char* logPath, size_t logSize)
+{
+    if (logPath)
+    {
+        karere::gLogger.logToFile(logPath, logSize);
+    }
+    services_init(postFunc, options);
+}
+
+void globalCleanup()
+{
+#ifndef KARERE_DISABLE_WEBRTC
+    rtcModule::globalCleanup();
+#endif
+    services_shutdown();
+}
 
 void RemoteLogger::log(krLogLevel level, const char* msg, size_t len, unsigned flags)
 {
@@ -40,4 +65,28 @@ void RemoteLogger::log(krLogLevel level, const char* msg, size_t len, unsigned f
             return err;
         });
 }
+
+#ifdef USE_LIBWEBSOCKETS
+
+void init_uv_timer(void *ctx, uv_timer_t *timer)
+{
+    uv_timer_init(((mega::LibuvWaiter *)(((megachat::MegaChatApiImpl *)ctx)->waiter))->eventloop, timer);
+}
+
+#else
+
+struct event_base *get_ev_loop(void *ctx)
+{
+    if (ctx)
+    {
+        return ((mega::LibeventWaiter *)(((megachat::MegaChatApiImpl *)ctx)->waiter))->eventloop;
+    }
+    else
+    {
+        return services_get_event_loop();
+    }
+}
+
+#endif
+
 }
