+++ src/net/libwsIO.cpp
@@ -24,24 +24,27 @@ void LibwsIO::addevents(::mega::Waiter* waiter, int)
     if (!initialized)
     {
         ::mega::LibeventWaiter *libeventWaiter = dynamic_cast<::mega::LibeventWaiter *>(waiter);
-        if (!libeventWaiter)
-        {
-            // Currently,LibwsIO is not compatible with waiters different than LibeventWaiter
-            exit(0);
-        }
-        
-        ws_global_init(&wscontext, libeventWaiter->eventloop, NULL,
+        ws_global_init(&wscontext, libeventWaiter ? libeventWaiter->eventloop : services_get_event_loop(), NULL,
         [](struct bufferevent* bev, void* userp)
         {
-            ws_read_callback(bev, userp);
+            karere::marshallCall([bev, userp]()
+            {
+                ws_read_callback(bev, userp);
+            }, NULL);
         },
         [](struct bufferevent* bev, short events, void* userp)
         {
-            ws_event_callback(bev, events, userp);
+            karere::marshallCall([bev, events, userp]()
+            {
+                ws_event_callback(bev, events, userp);
+            }, NULL);
         },
         [](int fd, short events, void* userp)
         {
-            ws_handle_marshall_timer_cb(0, events, userp);
+            karere::marshallCall([events, userp]()
+            {
+                ws_handle_marshall_timer_cb(0, events, userp);
+            }, NULL);
         });
         //ws_set_log_level(LIBWS_TRACE);
         initialized = true;
@@ -50,6 +53,11 @@ void LibwsIO::addevents(::mega::Waiter* waiter, int)
 
 WebsocketsClientImpl *LibwsIO::wsConnect(const char *ip, const char *host, int port, const char *path, bool ssl, WebsocketsClient *client)
 {
+    if (!initialized)
+    {
+        addevents(NULL, 0);
+    }
+
     int result;
     LibwsClient *libwsClient = new LibwsClient(mutex, client, appCtx);
